+++ packages/api-services/src/types/account.ts
@@ -1,0 +1,58 @@
+/**
+ * AccountCreateParams interface
+ */
+
+export interface AccountCreateParams {
+  /**
+   * Username field for the account
+   */
+  username: string;
+  /**
+   * 	Email field for the account
+   */
+  email: string;
+  /**
+   * 	Password field for the account
+   */
+  password: string;
+}
+
+export interface AccountVerifyParams {
+  /**
+   * Validation code sent by email
+   */
+  validation_code: string;
+}
+
+export interface AccountRecoverParams {
+  /**
+   * 	Email field for the account
+   */
+  email: string;
+}
+
+export interface AccountResetParams extends AccountVerifyParams {
+  /**
+   * 	Password field for the account
+   */
+  password: string;
+}
+
+export interface AccountSearchParams {
+  /**
+   * Username field for the account
+   */
+  username?: string;
+  /**
+   * 	Email field for the account
+   */
+  email?: string;
+  /**
+   * Ext ID field for the account (used in combination with provider)
+   */
+  ext_id?: string;
+  /**
+   * 	Provider field for the account (used in combination with ext_id)
+   */
+  provider?: string;
+}