+++ packages/react-ui/src/components/AccountReset/index.tsx
@@ -1,0 +1,4 @@
+import AccountReset, {AccountResetProps} from './AccountReset';
+
+export default AccountReset;
+export {AccountResetProps};