+++ packages/react-ui/src/components/AccountSignIn/index.tsx
@@ -1,4 +1,0 @@
-import AccountSignIn, {AccountSignInProps} from './AccountSignIn';
-
-export default AccountSignIn;
-export {AccountSignInProps};